Views and base Controllers updated.
Also it contains tasks, which are actually obsolete. Given the fact
that I will only use one other thread to make sql consults.
The Register class is for any kind of sql row.
Many classes such as Tasks and the service are useless. And will be
removed in the upcoming commitments.